requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ML / Vector search
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0

# PDF parsing
pdfplumber>=0.11.0

# LLM -- Groq (free tier, Llama 3.3 70B)
groq>=0.9.0

# Environment variable loading
python-dotenv>=1.0.0

# Dev / testing (optional)
# httpx>=0.27.0
# pytest>=8.0.0
# pytest-asyncio>=0.23.0